feat: add new test case
